+++ src/bask_project.c
@@ -28,7 +28,7 @@ void project_free_ll (struct bask_project** first)
 	Description: Inserts a new project with data.
 	InitVersion: 0.0.1
 */
-int project_insert (struct bask_project** first, short t_active, short t_state, char* t_project)
+short project_insert (struct bask_project** first, short t_active, short t_state, char* t_project)
 {
 	struct bask_project *newobj = malloc (sizeof (struct bask_project)), *preobj;
